+++ test/app/components/public_method_component.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class PublicMethodComponent < ActionView::Component::Base
+  def initialize(*); end
+
+  def public_method; end
+end